Move _check_fully_annotated to smile_inchi_inchikey_conversions.py
